i18nextract: sort the file locations.
Previously the order of the entries in the pot could jump randomly between
runs because the sort key was not stable (find_files returns the files in
random order, they're then added to the catalog location list in the same
order).

Now the order is deterministic, although it is not lexicographical (.py files
will come before .zcml files will come before .pt files; within each group the
files will be sorted).